tableau_streamlit_groups_monitoring/requirements.txt
tableauserverclient
streamlit
pandas
python-dotenv
